Moved logdet to math
